// src/contracts/newport/typechain/OptionMarket.ts
export declare namespace OptionMarket {
  export type OptionMarketParametersStruct = {
    maxBoardExpiry: PromiseOrValue<BigNumberish>;
    securityModule: PromiseOrValue<string>;
    feePortionReserved: PromiseOrValue<BigNumberish>;
    staticBaseSettlementFee: PromiseOrValue<BigNumberish>;
  };

  export type OptionMarketParametersStructOutput = [
    BigNumber,
    string,
    BigNumber,
    BigNumber
  ] & {
    maxBoardExpiry: BigNumber;
    securityModule: string;
    feePortionReserved: BigNumber;
    staticBaseSettlementFee: BigNumber;
  };

  export type TradeEventDataStruct = {
    expiry: PromiseOrValue<BigNumberish>;
    strikePrice: PromiseOrValue<BigNumberish>;
    optionType: PromiseOrValue<BigNumberish>;
    tradeDirection: PromiseOrValue<BigNumberish>;
    amount: PromiseOrValue<BigNumberish>;
    setCollateralTo: PromiseOrValue<BigNumberish>;
    isForceClose: PromiseOrValue<boolean>;
    spotPrice: PromiseOrValue<BigNumberish>;
    reservedFee: PromiseOrValue<BigNumberish>;
    totalCost: PromiseOrValue<BigNumberish>;
  };

  export type TradeEventDataStructOutput = [
    BigNumber,
    BigNumber,
    number,
    number,
    BigNumber,
    BigNumber,
    boolean,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    expiry: BigNumber;
    strikePrice: BigNumber;
    optionType: number;
    tradeDirection: number;
    amount: BigNumber;
    setCollateralTo: BigNumber;
    isForceClose: boolean;
    spotPrice: BigNumber;
    reservedFee: BigNumber;
    totalCost: BigNumber;
  };

  export type LiquidationEventDataStruct = {
    rewardBeneficiary: PromiseOrValue<string>;
    caller: PromiseOrValue<string>;
    returnCollateral: PromiseOrValue<BigNumberish>;
    lpPremiums: PromiseOrValue<BigNumberish>;
    lpFee: PromiseOrValue<BigNumberish>;
    liquidatorFee: PromiseOrValue<BigNumberish>;
    smFee: PromiseOrValue<BigNumberish>;
    insolventAmount: PromiseOrValue<BigNumberish>;
  };

  export type LiquidationEventDataStructOutput = [
    string,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    rewardBeneficiary: string;
    caller: string;
    returnCollateral: BigNumber;
    lpPremiums: BigNumber;
    lpFee: BigNumber;
    liquidatorFee: BigNumber;
    smFee: BigNumber;
    insolventAmount: BigNumber;
  };

  export type TradeInputParametersStruct = {
    strikeId: PromiseOrValue<BigNumberish>;
    positionId: PromiseOrValue<BigNumberish>;
    iterations: PromiseOrValue<BigNumberish>;
    optionType: PromiseOrValue<BigNumberish>;
    amount: PromiseOrValue<BigNumberish>;
    setCollateralTo: PromiseOrValue<BigNumberish>;
    minTotalCost: PromiseOrValue<BigNumberish>;
    maxTotalCost: PromiseOrValue<BigNumberish>;
  };

  export type TradeInputParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    number,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    strikeId: BigNumber;
    positionId: BigNumber;
    iterations: BigNumber;
    optionType: number;
    amount: BigNumber;
    setCollateralTo: BigNumber;
    minTotalCost: BigNumber;
    maxTotalCost: BigNumber;
  };

  export type ResultStruct = {
    positionId: PromiseOrValue<BigNumberish>;
    totalCost: PromiseOrValue<BigNumberish>;
    totalFee: PromiseOrValue<BigNumberish>;
  };

  export type ResultStructOutput = [BigNumber, BigNumber, BigNumber] & {
    positionId: BigNumber;
    totalCost: BigNumber;
    totalFee: BigNumber;
  };

  export type OptionBoardStruct = {
    id: PromiseOrValue<BigNumberish>;
    expiry: PromiseOrValue<BigNumberish>;
    iv: PromiseOrValue<BigNumberish>;
    frozen: PromiseOrValue<boolean>;
    strikeIds: PromiseOrValue<BigNumberish>[];
  };

  export type OptionBoardStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    boolean,
    BigNumber[]
  ] & {
    id: BigNumber;
    expiry: BigNumber;
    iv: BigNumber;
    frozen: boolean;
    strikeIds: BigNumber[];
  };

  export type StrikeStruct = {
    id: PromiseOrValue<BigNumberish>;
    strikePrice: PromiseOrValue<BigNumberish>;
    skew: PromiseOrValue<BigNumberish>;
    longCall: PromiseOrValue<BigNumberish>;
    shortCallBase: PromiseOrValue<BigNumberish>;
    shortCallQuote: PromiseOrValue<BigNumberish>;
    longPut: PromiseOrValue<BigNumberish>;
    shortPut: PromiseOrValue<BigNumberish>;
    boardId: PromiseOrValue<BigNumberish>;
  };

  export type StrikeStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    id: BigNumber;
    strikePrice: BigNumber;
    skew: BigNumber;
    longCall: BigNumber;
    shortCallBase: BigNumber;
    shortCallQuote: BigNumber;
    longPut: BigNumber;
    shortPut: BigNumber;
    boardId: BigNumber;
  };
}

export declare namespace OptionMarketPricer {
  export type VegaUtilFeeComponentsStruct = {
    preTradeAmmNetStdVega: PromiseOrValue<BigNumberish>;
    postTradeAmmNetStdVega: PromiseOrValue<BigNumberish>;
    vegaUtil: PromiseOrValue<BigNumberish>;
    volTraded: PromiseOrValue<BigNumberish>;
    NAV: PromiseOrValue<BigNumberish>;
    vegaUtilFee: PromiseOrValue<BigNumberish>;
  };

  export type VegaUtilFeeComponentsStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    preTradeAmmNetStdVega: BigNumber;
    postTradeAmmNetStdVega: BigNumber;
    vegaUtil: BigNumber;
    volTraded: BigNumber;
    NAV: BigNumber;
    vegaUtilFee: BigNumber;
  };

  export type VarianceFeeComponentsStruct = {
    varianceFeeCoefficient: PromiseOrValue<BigNumberish>;
    vega: PromiseOrValue<BigNumberish>;
    vegaCoefficient: PromiseOrValue<BigNumberish>;
    skew: PromiseOrValue<BigNumberish>;
    skewCoefficient: PromiseOrValue<BigNumberish>;
    ivVariance: PromiseOrValue<BigNumberish>;
    ivVarianceCoefficient: PromiseOrValue<BigNumberish>;
    varianceFee: PromiseOrValue<BigNumberish>;
  };

  export type VarianceFeeComponentsStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    varianceFeeCoefficient: BigNumber;
    vega: BigNumber;
    vegaCoefficient: BigNumber;
    skew: BigNumber;
    skewCoefficient: BigNumber;
    ivVariance: BigNumber;
    ivVarianceCoefficient: BigNumber;
    varianceFee: BigNumber;
  };

  export type TradeResultStruct = {
    amount: PromiseOrValue<BigNumberish>;
    premium: PromiseOrValue<BigNumberish>;
    optionPriceFee: PromiseOrValue<BigNumberish>;
    spotPriceFee: PromiseOrValue<BigNumberish>;
    vegaUtilFee: OptionMarketPricer.VegaUtilFeeComponentsStruct;
    varianceFee: OptionMarketPricer.VarianceFeeComponentsStruct;
    totalFee: PromiseOrValue<BigNumberish>;
    totalCost: PromiseOrValue<BigNumberish>;
    volTraded: PromiseOrValue<BigNumberish>;
    newBaseIv: PromiseOrValue<BigNumberish>;
    newSkew: PromiseOrValue<BigNumberish>;
  };

  export type TradeResultStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    OptionMarketPricer.VegaUtilFeeComponentsStructOutput,
    OptionMarketPricer.VarianceFeeComponentsStructOutput,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    amount: BigNumber;
    premium: BigNumber;
    optionPriceFee: BigNumber;
    spotPriceFee: BigNumber;
    vegaUtilFee: OptionMarketPricer.VegaUtilFeeComponentsStructOutput;
    varianceFee: OptionMarketPricer.VarianceFeeComponentsStructOutput;
    totalFee: BigNumber;
    totalCost: BigNumber;
    volTraded: BigNumber;
    newBaseIv: BigNumber;
    newSkew: BigNumber;
  };
}
// ...
export interface TradeEventObject {
  trader: string;
  strikeId: BigNumber;
  positionId: BigNumber;
  trade: OptionMarket.TradeEventDataStructOutput;
  tradeResults: OptionMarketPricer.TradeResultStructOutput[];
  liquidation: OptionMarket.LiquidationEventDataStructOutput;
  longScaleFactor: BigNumber;
  timestamp: BigNumber;
}
export type TradeEvent = TypedEvent<
  [
    string,
    BigNumber,
    BigNumber,
    OptionMarket.TradeEventDataStructOutput,
    OptionMarketPricer.TradeResultStructOutput[],
    OptionMarket.LiquidationEventDataStructOutput,
    BigNumber,
    BigNumber
  ],
  TradeEventObject
>;

export type TradeEventFilter = TypedEventFilter<TradeEvent>;

export interface OptionMarket extends BaseContract {
// ...
  filters: {
// ...
      liquidation?: null,
      longScaleFactor?: null,
      timestamp?: null
    ): TradeEventFilter;
    Trade(
      trader?: PromiseOrValue<string> | null,
      strikeId?: PromiseOrValue<BigNumberish> | null,
      positionId?: PromiseOrValue<BigNumberish> | null,
      trade?: null,
      tradeResults?: null,
      liquidation?: null,
      longScaleFactor?: null,
      timestamp?: null
    ): TradeEventFilter;
  };
// ...
}